use Long.MAX_VALUE instead of Integer.MAX_VALUE

+++ com.sap.sailing.server/src/com/sap/sailing/server/ModeratorApp.java
@@ -336,7 +336,7 @@ public class ModeratorApp extends Servlet {
                 jsonRace.put("name", race.getName());
                 jsonRace.put("boatclass", race.getBoatClass()==null?"":race.getBoatClass().getName());
                 TimePoint start = getService().getDomainFactory().trackEvent(event).getTrackedRace(race).getStart();
-                jsonRace.put("start", start==null?Integer.MAX_VALUE:start.asMillis());
+                jsonRace.put("start", start==null?Long.MAX_VALUE:start.asMillis());
                 JSONArray jsonLegs = new JSONArray();
                 for (Leg leg : race.getCourse().getLegs()) {
                     JSONObject jsonLeg = new JSONObject();
